Migrate dbmanager to separate package
