+++ charts/otel-shopping-cart/templates/user-service.yaml
@@ -2,10 +2,12 @@ kind: Service
 apiVersion: v1
 metadata:
   name: {{ .Values.user.serviceName }}
-  app: user
+  labels:
+    app: user
 spec:
   selector:
     app: {{ .Values.user.serviceName }}
   ports:
     - port: {{ .Values.user.port }}
+      name: http
       targetPort: {{ .Values.user.port }}
