+++ cmd/interrupter/main.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/trstringer/otel-shopping-cart/pkg/users"
+	"github.com/trstringer/otel-shopping-cart/pkg/dbmanager"
 )
 
 var (
@@ -23,7 +23,7 @@ var rootCmd = &cobra.Command{
 	Short: "Service interrupter",
 	Long:  `Interrupt service and cause quality issues.`,
 	Run: func(cmd *cobra.Command, args []string) {
-		dbm := users.NewDBManager(dbSQLAddress, "otel_shopping_cart", dbSQLUser, os.Getenv("DB_PASSWORD"))
+		dbm := dbmanager.NewDBManager(dbSQLAddress, "otel_shopping_cart", dbSQLUser, os.Getenv("DB_PASSWORD"))
 		users, err := dbm.GetAllUsers()
 		if err != nil {
 			fmt.Printf("Error getting users: %v\n", err)
