+++ pkg/dbmanager/metrics.go
@@ -1,4 +1,4 @@
-package cart
+package dbmanager
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
