--- pkg/users/metrics.go
@@ -1,13 +1,0 @@
-package users
-
-import (
-	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/promauto"
-)
-
-var (
-	dbmanagerErrors = promauto.NewCounter(prometheus.CounterOpts{
-		Name: "dbmanager_error",
-		Help: "dbmanager error count",
-	})
-)
